Add error handling
isQuadratic function is used to see if 'a' is equal to '0'.
validateValue function checks if the value is a number.
validateDataFormat function checks if the format of a file
is right(3 numbers separated by a single space, LF/CRLF in the end)
promptValue and readValue are used to read values from stdin.